requirements.txt
Django~=3.2
djangorestframework~=3.12.4
django-rest-swagger~=2.2.0
django-rest-auth~=0.9.5
drf-writable-nested~=0.6.3
djoser~=2.1.0
cryptography~=3.4.7
pip~=21.1.3
pytz~=2021.1
sqlparse~=0.4.1
psycopg2~=2.8.6
asgiref~=3.3.4
PyYAML~=5.4.1
Jinja2~=3.0.1
MarkupSafe~=2.0.1
certifi~=2021.5.30
chardet~=4.0.0
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
requests~=2.25.1
uritemplate~=3.0.1
six~=1.16.0
oauthlib~=3.1.1
idna~=2.10
urllib3~=1.26.6
simplejson~=3.17.2
pycparser~=2.20
setuptools~=56.0.0
PyJWT~=2.1.0
Pillow~=9.1.1
